+++ src/__tests__/Link.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Link from './Components/Link/Link';
+import Link from '../Components/Link/Link';
 
 it('Link should render correctly', () => {
     const tree = renderer
@@ -9,6 +9,7 @@ it('Link should render correctly', () => {
         .toJSON();
     expect(tree).toMatchSnapshot();
 });
+
 it('Link with href attribute should render correctly', () => {
     const tree = renderer
         .create(<Link href="se lascar" />)
